requirements.txt
python-dateutil==2.8.2
chalice==1.24.0
